# Docs/Api/cpay-v2-openapi.yaml
openapi: 3.0.3
info:
  title: CPay Gateway API
  version: 1.6.0
  description: Merchant payment, callback operations, reconciliation, finance, and operations endpoints for CPay.
security:
  - CPayV2Signature: []
paths:
...
  /api/v2/payments/{reference}:
    get:
      summary: Check transaction status
      operationId: transactionStatus
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
        - name: merchantNumber
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction status
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v2/accounts/validate:
    post:
      summary: Validate an account name before payout
      operationId: validateAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountValidationRequest'
      responses:
        '200':
          description: Account validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountValidationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v2/checkout/{token}/pay:
    post:
      summary: Submit a hosted checkout payment
      operationId: payHostedCheckout
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedCheckoutPayRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/HostedCheckoutPayRequest'
      responses:
        '202':
          description: Hosted checkout accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResult'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v2/statements:
    get:
      summary: Export merchant statement rows
      operationId: exportMerchantStatements
      parameters:
        - name: merchantNumber
          in: query
          required: true
          schema:
            type: string
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [json, csv]
            default: json
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5000
      responses:
        '200':
          description: Merchant statement export
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatementExportResponse'
            text/csv:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v2/admin/treasury/positions:
    get:
      summary: List active treasury positions with net-available balances
      operationId: listTreasuryPositions
      security:
        - AdminBasicAuth: []
      responses:
        '200':
          description: Active treasury positions (available, reserved, net-available)
  /api/v2/admin/treasury/positions/{currency}:
    get:
      summary: Get one treasury position by ISO currency code
      operationId: getTreasuryPosition
      security:
        - AdminBasicAuth: []
      parameters:
        - name: currency
          in: path
          required: true
          schema:
            type: string
            pattern: '^[A-Z]{3}$'
      responses:
        '200':
          description: Treasury position for the currency
        '404':
          description: No active position for the currency
        '400':
          $ref: '#/components/responses/BadRequest'
components:
  securitySchemes:
    CPayV2Signature:
      type: apiKey
      in: header
      name: X-CPay-Signature
    AdminBasicAuth:
      type: http
      scheme: basic
...
  schemas:
...
    PaymentResult:
      type: object
      properties:
        reference:
          type: string
        transactionId:
          type: string
        status:
          type: string
        channel:
          type: string
        currency:
          type: string
        message:
          type: string
        providerResponse:
          type: string
    AccountValidationRequest:
      type: object
      required: [merchantNumber, msisdn]
      properties:
        merchantNumber:
          type: string
        msisdn:
          type: string
    AccountValidationResponse:
      type: object
      properties:
        msisdn:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        name:
          type: string
        status:
          type: string
...
    HostedCheckoutPayRequest:
      type: object
      required: [payerAccount]
      properties:
        payerAccount:
          type: string
        channel:
          type: string
    StatementExportResponse:
      type: object
      properties:
        merchantNumber:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        count:
          type: integer
        rows:
          type: array
          items:
            $ref: '#/components/schemas/StatementRow'
    StatementRow:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdOn:
          type: string
        gatewayId:
          type: string
        transactionType:
          type: string
        amount:
          type: number
        currency:
          type: string
        merchantReference:
          type: string
        transactionId:
          type: string
        transactionStatus:
          type: string
...
  responses:
    BadRequest:
      description: Request was rejected
    Unauthorized:
      description: Request signature or credentials were rejected