# InitializrSpringbootProjectFresh/src/main/java/net/citotech/cito/api/v2/PaymentsV2Controller.java
package net.citotech.cito.api.v2;

import com.fasterxml.jackson.databind.ObjectMapper;
import jakarta.servlet.http.HttpServletRequest;
import java.util.List;
import java.util.Optional;
import java.util.UUID;
import net.citotech.cito.Model.Balance;
import net.citotech.cito.Model.Merchant;
import net.citotech.cito.PaymentOrchestrationService;
import net.citotech.cito.api.v2.dto.AccountValidationRequest;
import net.citotech.cito.api.v2.dto.ApiErrorResponse;
import net.citotech.cito.api.v2.dto.PaymentChannelResponse;
import net.citotech.cito.api.v2.dto.PaymentRequest;
import net.citotech.cito.api.v2.dto.PaymentResult;
import net.citotech.cito.api.v2.dto.PaymentStatusResponse;
import net.citotech.cito.api.v2.dto.StatementExportResponse;
import net.citotech.cito.export.TabularExportService;
import net.citotech.cito.gateway.PaymentGatewayException;
import net.citotech.cito.payout.PayoutControlService;
import net.citotech.cito.payout.PayoutControlService.PayoutEvaluation;
import org.springframework.http.HttpHeaders;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping(path = "/api/v2")
public class PaymentsV2Controller {
    private final PaymentOrchestrationService paymentOrchestrationService;
    private final PaymentStatusService paymentStatusService;
    private final V2RequestSecurityService securityService;
    private final IdempotencyService idempotencyService;
    private final AccountValidationService accountValidationService;
    private final MerchantStatementExportService statementExportService;
    private final PayoutControlService payoutControlService;
    private final ObjectMapper objectMapper;

    public PaymentsV2Controller(
            PaymentOrchestrationService paymentOrchestrationService,
            PaymentStatusService paymentStatusService,
            V2RequestSecurityService securityService,
            IdempotencyService idempotencyService,
            AccountValidationService accountValidationService,
            MerchantStatementExportService statementExportService,
            PayoutControlService payoutControlService,
            ObjectMapper objectMapper) {
        this.paymentOrchestrationService = paymentOrchestrationService;
        this.paymentStatusService = paymentStatusService;
        this.securityService = securityService;
        this.idempotencyService = idempotencyService;
        this.accountValidationService = accountValidationService;
        this.statementExportService = statementExportService;
        this.payoutControlService = payoutControlService;
        this.objectMapper = objectMapper;
    }

    @PostMapping(path = "/payments/collect")
    public ResponseEntity<?> collect(@RequestBody String body, HttpServletRequest servletRequest) {
        try {
            PaymentRequest request = objectMapper.readValue(body, PaymentRequest.class);
            Merchant merchant =
                    securityService.verify(servletRequest, body, request.getMerchantNumber());
            String idempotencyKey = servletRequest.getHeader("X-CPay-Idempotency-Key");
            Optional<PaymentResult> existing =
                    idempotencyService.findExisting(
                            request.getMerchantNumber(), idempotencyKey, body);
            if (existing.isPresent()) {
                return ResponseEntity.ok(existing.get());
            }
            PaymentResult result =
                    paymentOrchestrationService.collect(
                            request, merchant, servletRequest.getRemoteAddr());
            idempotencyService.record(request.getMerchantNumber(), idempotencyKey, body, result);
            return ResponseEntity.accepted().body(result);
        } catch (V2RequestSecurityException e) {
            return error(HttpStatus.UNAUTHORIZED, "UNAUTHORIZED", e.getMessage());
        } catch (PaymentGatewayException e) {
            return error(HttpStatus.BAD_REQUEST, "PAYMENT_REJECTED", e.getMessage());
        } catch (Exception e) {
            return error(HttpStatus.BAD_REQUEST, "INVALID_REQUEST", "Invalid collect request");
        }
    }

    @PostMapping(path = "/payments/payout")
    public ResponseEntity<?> payout(@RequestBody String body, HttpServletRequest servletRequest) {
        try {
            PaymentRequest request = objectMapper.readValue(body, PaymentRequest.class);
            Merchant merchant =
                    securityService.verify(servletRequest, body, request.getMerchantNumber());
            String idempotencyKey = servletRequest.getHeader("X-CPay-Idempotency-Key");
            Optional<PaymentResult> existing =
                    idempotencyService.findExisting(
                            request.getMerchantNumber(), idempotencyKey, body);
            if (existing.isPresent()) {
                return ResponseEntity.ok(existing.get());
            }
            // Payout risk controls: a configured limit breach or a review-required trigger parks
            // the payout in the maker-checker approval queue instead of executing it. With no
            // control row (or a disabled one) this returns EXECUTE and behavior is unchanged.
            PayoutEvaluation control = payoutControlService.evaluate(request, merchant, "system");
            if (control.isApprovalRequired()) {
                PaymentResult pending = new PaymentResult();
                pending.setReference(request.getReference());
                pending.setStatus("APPROVAL_PENDING");
                pending.setChannel(request.getChannel());
                pending.setCurrency(request.getCurrency());
                pending.setMessage(
                        "Payout requires maker-checker approval: " + control.reasonCode());
                idempotencyService.record(
                        request.getMerchantNumber(), idempotencyKey, body, pending);
                return ResponseEntity.status(HttpStatus.ACCEPTED).body(pending);
            }
            PaymentResult result =
                    paymentOrchestrationService.payout(
                            request, merchant, servletRequest.getRemoteAddr());
            idempotencyService.record(request.getMerchantNumber(), idempotencyKey, body, result);
            return ResponseEntity.accepted().body(result);
        } catch (V2RequestSecurityException e) {
            return error(HttpStatus.UNAUTHORIZED, "UNAUTHORIZED", e.getMessage());
        } catch (PaymentGatewayException e) {
            return error(HttpStatus.BAD_REQUEST, "PAYMENT_REJECTED", e.getMessage());
        } catch (Exception e) {
            return error(HttpStatus.BAD_REQUEST, "INVALID_REQUEST", "Invalid payout request");
        }
    }

    @GetMapping(path = "/channels")
    public List<PaymentChannelResponse> channels() {
        return paymentOrchestrationService.listChannels();
    }

    @GetMapping(path = "/balances")
    public ResponseEntity<?> balances(
            @RequestParam("merchantNumber") String merchantNumber,
            HttpServletRequest servletRequest) {
        try {
            Merchant merchant = securityService.verify(servletRequest, "", merchantNumber);
            List<Balance> balances = paymentOrchestrationService.balances(merchantNumber, merchant);
            return ResponseEntity.ok(balances);
        } catch (V2RequestSecurityException e) {
            return error(HttpStatus.UNAUTHORIZED, "UNAUTHORIZED", e.getMessage());
        } catch (PaymentGatewayException e) {
            return error(HttpStatus.BAD_REQUEST, "BALANCE_REJECTED", e.getMessage());
        }
    }

    @GetMapping(path = "/payments/{reference}")
    public ResponseEntity<?> status(
            @PathVariable("reference") String reference,
            @RequestParam("merchantNumber") String merchantNumber,
            HttpServletRequest servletRequest) {
        try {
            Merchant merchant = securityService.verify(servletRequest, "", merchantNumber);
            PaymentStatusResponse status = paymentStatusService.getStatus(merchant, reference);
            return ResponseEntity.ok(status);
        } catch (V2RequestSecurityException e) {
            return error(HttpStatus.UNAUTHORIZED, "UNAUTHORIZED", e.getMessage());
        } catch (PaymentGatewayException e) {
            return error(HttpStatus.BAD_REQUEST, "STATUS_REJECTED", e.getMessage());
        }
    }

    @PostMapping(path = "/accounts/validate")
    public ResponseEntity<?> validateAccount(
            @RequestBody String body, HttpServletRequest servletRequest) {
        try {
            AccountValidationRequest request =
                    objectMapper.readValue(body, AccountValidationRequest.class);
            Merchant merchant =
                    securityService.verify(servletRequest, body, request.getMerchantNumber());
            return ResponseEntity.ok(accountValidationService.validate(request, merchant));
        } catch (V2RequestSecurityException e) {
            return error(HttpStatus.UNAUTHORIZED, "UNAUTHORIZED", e.getMessage());
        } catch (PaymentGatewayException e) {
            return error(HttpStatus.BAD_REQUEST, "ACCOUNT_VALIDATION_REJECTED", e.getMessage());
        } catch (Exception e) {
            return error(
                    HttpStatus.BAD_REQUEST,
                    "INVALID_REQUEST",
                    "Invalid account validation request");
        }
    }

    @GetMapping(path = "/statements")
    public ResponseEntity<?> statements(
            @RequestParam("merchantNumber") String merchantNumber,
            @RequestParam("startDate") String startDate,
            @RequestParam("endDate") String endDate,
            @RequestParam(value = "format", defaultValue = "json") String format,
            @RequestParam(value = "limit", required = false) Integer limit,
            @RequestParam(value = "cursor", required = false) String cursor,
            HttpServletRequest servletRequest) {
        try {
            Merchant merchant = securityService.verify(servletRequest, "", merchantNumber);
            StatementExportResponse export =
                    statementExportService.export(
                            merchant, merchantNumber, startDate, endDate, limit, cursor);
            if ("csv".equalsIgnoreCase(format)) {
                String filename =
                        "cpay-statement-"
                                + merchantNumber
                                + "-"
                                + startDate
                                + "-to-"
                                + endDate
                                + ".csv";
                return ResponseEntity.ok()
                        .header(
                                HttpHeaders.CONTENT_DISPOSITION,
                                "attachment; filename=\"" + filename + "\"")
                        .contentType(
                                MediaType.parseMediaType(TabularExportService.CSV_CONTENT_TYPE))
                        .body(statementExportService.toCsv(export));
            }
            if ("xlsx".equalsIgnoreCase(format)) {
                String filename =
                        "cpay-statement-"
                                + merchantNumber
                                + "-"
                                + startDate
                                + "-to-"
                                + endDate
                                + ".xlsx";
                return ResponseEntity.ok()
                        .header(
                                HttpHeaders.CONTENT_DISPOSITION,
                                "attachment; filename=\"" + filename + "\"")
                        .contentType(
                                MediaType.parseMediaType(TabularExportService.XLSX_CONTENT_TYPE))
                        .body(statementExportService.toXlsx(export));
            }
            return ResponseEntity.ok(export);
        } catch (V2RequestSecurityException e) {
            return error(HttpStatus.UNAUTHORIZED, "UNAUTHORIZED", e.getMessage());
        } catch (PaymentGatewayException e) {
            return error(HttpStatus.BAD_REQUEST, "STATEMENT_EXPORT_REJECTED", e.getMessage());
        }
    }

    private ResponseEntity<ApiErrorResponse> error(HttpStatus status, String code, String message) {
        return ResponseEntity.status(status)
                .body(new ApiErrorResponse(code, message, UUID.randomUUID().toString()));
    }
}